+++ src/api/networks.rs
@@ -1,10 +1,56 @@
-use crate::{api::ApiResource, Result};
+use crate::{api::ApiResource, models, opts, util::url, Result};
 
 impl_api_ty!(
     Network => name
 );
 
 impl<'podman> Network<'podman> {
+    api_doc! {
+    Network => DeleteLibpod
+    /// Delete this container. To delete this network forcefully use
+    /// [`Network::force_delete`](Network::force_delete).
+    ///
+    /// Examples:
+    ///
+    /// ```no_run
+    /// let podman = Podman::unix("/run/user/1000/podman/podman.sock");
+    ///
+    /// match podman.networks().get("some-network").delete().await {
+    ///     Ok(info) => println!("{:?}", info),
+    ///     Err(e) => eprintln!("{}", e),
+    /// }
+    /// ```
+    |
+    pub async fn delete(&self) -> Result<models::NetworkRmReport> {
+        self.podman
+            .delete_json(&format!("/libpod/networks/{}", &self.name))
+            .await
+    }}
+
+    api_doc! {
+    Network => DeleteLibpod
+    /// Force remove this network removing associated containers. To delete network normally use
+    /// [`Network::delete`](Network::delete).
+    ///
+    /// Examples:
+    ///
+    /// ```no_run
+    /// let podman = Podman::unix("/run/user/1000/podman/podman.sock");
+    ///
+    /// match podman.networks().get("some-network").force_delete().await {
+    ///     Ok(info) => println!("{:?}", info),
+    ///     Err(e) => eprintln!("{}", e),
+    /// }
+    /// ```
+    |
+    pub async fn force_delete(&self) -> Result<models::NetworkRmReport> {
+        let ep = url::construct_ep(
+            &format!("/libpod/networks/{}", &self.name),
+            Some(url::encoded_pair("force", true.to_string())),
+        );
+        self.podman.delete_json(&ep).await
+    }}
+
     api_doc! {
     Network => ExistsLibpod
     /// Quick way to determine if a network exists by name or id.